refactor: データフェッチをSuspense境界とuse()に移行 (#86)
- AppLayout を Server Component 化し、profiles(xp, level) と
  開発者判定の promise を await せずに開始して Client 側へ渡す
  （既存のクライアント実装は AppLayoutClient に分離）
- XpBadge: useEffect + useState での取得を Suspense + use() に移行。
  sessionStorage キャッシュは「遷移中の fallback 表示」用として存続
  （useSyncExternalStore で SSR 安全に参照）。TTL による再取得制御は
  サーバー側で毎リクエスト最新値を取得するため不要となり削除
- Sidebar: /api/developer/me への client fetch を廃止し、サーバーで
  isDeveloperUserId を直接判定した promise を Suspense + use() で表示
- XP 反映は Server Action の revalidatePath / 画面遷移で AppLayout が
  再レンダーされ新しい promise が渡ることで従来以上に即時反映される
